refactor: reorganize advanced_function_tests.rs into behavior-specific modules
- Split 614-line test file into focused submodules:
  - math_functions.rs: EXP, LN, LOG, SIGN, PI functions
  - trig_functions.rs: SIN, COS, TAN, ASIN, ACOS, ATAN, ATAN2, RADIANS, DEGREES
  - conditional_functions.rs: GREATEST, LEAST, IF functions
  - nested_expressions.rs: Complex expressions combining multiple functions
  - truncate_functions.rs: TRUNCATE function with various precisions
  - fixtures.rs: Common constants and helper functions
- Centralize common fixtures and constants
- Maintain executor suite success
- Improve test organization and maintainability

Closes #707

Amp-Thread-ID: https://ampcode.com/threads/T-be3c3b2c-b179-4c15-9566-45494eeab244
